Consolidate app_strings tests
